+++ modal/experimental/__init__.py
@@ -1,17 +1,20 @@
 # Copyright Modal Labs 2025
+import os
 from dataclasses import dataclass
-from typing import Any, Callable, Optional
+from pathlib import Path
+from typing import Literal, Optional, Union
 
 from modal_proto import api_pb2
 
 from .._clustered_functions import ClusterInfo, get_cluster_info as _get_cluster_info
-from .._functions import _Function
 from .._object import _get_environment_name
-from .._partial_function import _PartialFunction, _PartialFunctionFlags
+from .._partial_function import _clustered
 from .._runtime.container_io_manager import _ContainerIOManager
-from .._utils.async_utils import synchronizer
+from .._utils.async_utils import synchronize_api, synchronizer
 from ..client import _Client
 from ..exception import InvalidError
+from ..image import DockerfileSpec, ImageBuilderVersion, _Image, _ImageRegistryConfig
+from ..secret import _Secret
 
 
 def stop_fetching_inputs():
@@ -34,39 +37,11 @@ def set_local_input_concurrency(concurrency: int):
     _ContainerIOManager.set_input_concurrency(concurrency)
 
 
-def clustered(size: int, broadcast: bool = True):
-    """Provision clusters of colocated and networked containers for the Function.
-
-    Parameters:
-    size: int
-        Number of containers spun up to handle each input.
-    broadcast: bool = True
-        If True, inputs will be sent simultaneously to each container. Otherwise,
-        inputs will be sent only to the rank-0 container, which is responsible for
-        delegating to the workers.
-    """
-
-    assert broadcast, "broadcast=False has not been implemented yet!"
-
-    if size <= 0:
-        raise ValueError("cluster size must be greater than 0")
-
-    def wrapper(raw_f: Callable[..., Any]) -> _PartialFunction:
-        if isinstance(raw_f, _Function):
-            raw_f = raw_f.get_raw_f()
-            raise InvalidError(
-                f"Applying decorators for {raw_f} in the wrong order!\nUsage:\n\n"
-                "@app.function()\n@modal.clustered()\ndef clustered_function():\n    ..."
-            )
-        return _PartialFunction(
-            raw_f, _PartialFunctionFlags.FUNCTION | _PartialFunctionFlags.CLUSTERED, cluster_size=size
-        )
-
-    return wrapper
-
-
 def get_cluster_info() -> ClusterInfo:
     return _get_cluster_info()
+
+
+clustered = synchronize_api(_clustered, target_module=__name__)
 
 
 @dataclass
@@ -103,3 +78,82 @@ async def list_deployed_apps(environment_name: str = "", client: Optional[_Clien
                 )
             )
     return app_infos
+
+
+@synchronizer.create_blocking
+async def raw_dockerfile_image(
+    path: Union[str, Path],
+    force_build: bool = False,
+) -> _Image:
+    """
+    Build a Modal Image from a local Dockerfile recipe without any changes.
+
+    Unlike for `modal.Image.from_dockerfile`, the provided recipe will not be embellished with
+    steps to install dependencies for the Modal client package. As a consequence, the resulting
+    Image cannot be used with a modal Function unless those dependencies are added in a subsequent
+    layer. It _can_ be directly used with a modal Sandbox, which does not need the Modal client.
+
+    We expect to support this experimental function until the `2025.04` Modal Image Builder is
+    stable, at which point Modal Image recipes will no longer install the client dependencies
+    by default. At that point, users can upgrade their Image Builder Version and migrate to
+    `modal.Image.from_dockerfile` for usecases supported by this function.
+
+    """
+
+    def build_dockerfile(version: ImageBuilderVersion) -> DockerfileSpec:
+        with open(os.path.expanduser(path)) as f:
+            commands = f.read().split("\n")
+        return DockerfileSpec(commands=commands, context_files={})
+
+    return _Image._from_args(
+        dockerfile_function=build_dockerfile,
+        force_build=force_build,
+    )
+
+
+@synchronizer.create_blocking
+async def raw_registry_image(
+    tag: str,
+    registry_secret: Optional[_Secret] = None,
+    credential_type: Literal["static", "aws", "gcp", None] = None,
+    force_build: bool = False,
+) -> _Image:
+    """
+    Build a Modal Image from a public or private image registry without any changes.
+
+    Unlike for `modal.Image.from_registry`, the provided recipe will not be embellished with
+    steps to install dependencies for the Modal client package. As a consequence, the resulting
+    Image cannot be used with a modal Function unless those dependencies are added in a subsequent
+    layer. It _can_ be directly used with a modal Sandbox, which does not need the Modal client.
+
+    We expect to support this experimental function until the `2025.04` Modal Image Builder is
+    stable, at which point Modal Image recipes will no longer install the client dependencies
+    by default. At that point, users can upgrade their Image Builder Version and migrate to
+    `modal.Image.from_registry` for usecases supported by this function.
+
+    """
+
+    def build_dockerfile(version: ImageBuilderVersion) -> DockerfileSpec:
+        commands = [f"FROM {tag}"]
+        return DockerfileSpec(commands=commands, context_files={})
+
+    if registry_secret:
+        if credential_type is None:
+            raise InvalidError("credential_type must be provided when using a registry_secret")
+        elif credential_type == "static":
+            auth_type = api_pb2.REGISTRY_AUTH_TYPE_STATIC_CREDS
+        elif credential_type == "aws":
+            auth_type = api_pb2.REGISTRY_AUTH_TYPE_AWS
+        elif credential_type == "gcp":
+            auth_type = api_pb2.REGISTRY_AUTH_TYPE_GCP
+        else:
+            raise InvalidError(f"Invalid credential_type: {credential_type!r}")
+        registry_config = _ImageRegistryConfig(auth_type, registry_secret)
+    else:
+        registry_config = None
+
+    return _Image._from_args(
+        dockerfile_function=build_dockerfile,
+        image_registry_config=registry_config,
+        force_build=force_build,
+    )